Move remote queries test files to be under remote-queries dir
